Дождь кубов 2.2

+++ Assets/Scripts/interfaces/IView.cs
@@ -1,6 +1,6 @@
 ﻿public interface IView
 {
-    void UpdateSpawned(int count);
-    void UpdateInstantiated(int count);
-    void UpdateActive(int count);
+    void UpdateSpawned(ulong count);
+    void UpdateInstantiated(ulong count);
+    void UpdateActive(ulong count);
 }
